refactor(portfolio): replace relative imports with #/ alias
Convert all remaining ../ relative path imports to #/ alias imports across 11 files for consistency with project conventions.
