+++ projects/portfolio/src/client/components/chat/chat-settings/settings/auto-model-toggle.tsx
@@ -1,5 +1,5 @@
 import { ToggleInput } from '#/client/components/input/toggle-input'
-import { useChatSettingsContext } from '../chat-settings-context'
+import { useChatSettingsContext } from '#/client/components/chat/chat-settings/chat-settings-context'
 
 export function AutoModelToggle() {
   const { autoModel, handleToggleAutoModel } = useChatSettingsContext()
