+++ projects/portfolio/src/client/pages/diff/index.tsx
@@ -2,7 +2,7 @@ import Editor from '@monaco-editor/react'
 import type { ComponentType } from 'react'
 import { useEffect, useState } from 'react'
 import ReactDiffViewerModule from 'react-diff-viewer'
-import { loadDiffState, saveDiffState } from '../../storage/diff-state'
+import { loadDiffState, saveDiffState } from '#/client/storage/diff-state'
 
 const reactDiffViewerImport = ReactDiffViewerModule as ComponentType<any> | { default: ComponentType<any> }
 const ReactDiffViewer =
